API URLs and description changes for SAR Enhancements
