--- reInvent2017/geocode/geocode.yml
@@ -1,48 +1,0 @@
-AWSTemplateFormatVersion: '2010-09-09'
-Transform: 'AWS::Serverless-2016-10-31'
-
-Parameters:
-  HereAppId:
-    Type: String
-    Description: HERE Technologies API App ID
-  HereAppCode:
-    Type: String
-    Description: HERE Technologies API App Code
-
-Resources:
-  GeocodeFunction:
-    Type: 'AWS::Serverless::Function'
-    Properties:
-      Runtime: nodejs6.10
-      Handler: geocode.geocodeGET
-      Description: 'Convert street addresses to geo-coordinates and vice-versa with forward geocoding, including landmarks, and reverse geocoding.'
-      Environment:
-        Variables:
-          HERE_APP_ID: !Ref HereAppId
-          HERE_APP_CODE: !Ref HereAppCode
-      Events:
-        GET:
-          Type: Api
-          Properties:
-            Path: /geocode/{searchtext}
-            Method: get
-  GeocodeSuggestFunction:
-    Type: 'AWS::Serverless::Function'
-    Properties:
-      Runtime: nodejs6.10
-      Handler: geocodesuggest.geocodesuggestGET
-      Description: 'Correct misspellings and get better suggestions for your address search with fewer keystrokes.'
-      Environment:
-        Variables:
-          HERE_APP_ID: !Ref HereAppId
-          HERE_APP_CODE: !Ref HereAppCode
-      Events:
-        GET:
-          Type: Api
-          Properties:
-            Path: /geocodesuggest/{query}
-            Method: get
-Outputs:
-  HereApiURL:
-    Description: "AWS API Gateway URL for Here.com GeoCode & GeoCode Suggest (aka autocomplete) API (CIT)."
-    Value: !Sub "https://${ServerlessRestApi}.execute-api.${AWS::Region}.amazonaws.com/Stage/"
