--- reInvent2017/position/position.yml
@@ -1,32 +1,0 @@
-AWSTemplateFormatVersion: '2010-09-09'
-Transform: 'AWS::Serverless-2016-10-31'
-
-Parameters:
-  HereAppId:
-    Type: String
-    Description: HERE Maps API App ID (https://developer.here.com/faqs#access-control)
-  HereAppCode:
-    Type: String
-    Description: HERE Maps API App Code (https://developer.here.com/faqs#access-control)
-
-Resources:
-  PositionFunction:
-    Type: 'AWS::Serverless::Function'
-    Properties:
-      Runtime: nodejs6.10
-      Handler: position.positionPOST
-      Description: 'Provides positioning estimates based on global Wi-Fi and Cell coverage.'
-      Environment:
-        Variables:
-          HERE_APP_ID: !Ref HereAppId
-          HERE_APP_CODE: !Ref HereAppCode
-      Events:
-        GET:
-          Type: Api
-          Properties:
-            Path: /position
-            Method: post
-Outputs:
-  HereApiURL:
-    Description: "AWS API Gateway URL for Here.com Position API (CIT)."
-    Value: !Sub "https://${ServerlessRestApi}.execute-api.${AWS::Region}.amazonaws.com/Stage/"
