--- reInvent2017/routing/routing.yml
@@ -1,49 +1,0 @@
-AWSTemplateFormatVersion: '2010-09-09'
-Transform: 'AWS::Serverless-2016-10-31'
-
-Parameters:
-  HereAppId:
-    Type: String
-    Description: HERE Technologies API App ID
-  HereAppCode:
-    Type: String
-    Description: HERE Technologies API App Code
-
-Resources:
-  RoutingFunction:
-    Type: 'AWS::Serverless::Function'
-    Properties:
-      Runtime: nodejs6.10
-      Handler: routing.routingGET
-      Description: 'Provide precise instructions to a destination using various transport modes (e.g., car, truck, public transit, bicycle) and leveraging different algorithms (e.g., matrix, isoline routing).'
-      Environment:
-        Variables:
-          HERE_APP_ID: !Ref HereAppId
-          HERE_APP_CODE: !Ref HereAppCode
-      Events:
-        GET:
-          Type: Api
-          Properties:
-            Path: /routing
-            Method: get
-  WaypointSequenceFunction:
-    Type: 'AWS::Serverless::Function'
-    Properties:
-      Runtime: nodejs6.10
-      Handler: waypointseq.waypointseqGET
-      Description: 'Calculate a list of sequence-optimized waypoints.'
-      Environment:
-        Variables:
-          HERE_APP_ID: !Ref HereAppId
-          HERE_APP_CODE: !Ref HereAppCode
-      Events:
-        GET:
-          Type: Api
-          Properties:
-            Path: /waypointseq
-            Method: get
-
-Outputs:
-  HereApiURL:
-    Description: "AWS API Gateway URL for Here.com Routing & Waypoint Sequence API (CIT)."
-    Value: !Sub "https://${ServerlessRestApi}.execute-api.${AWS::Region}.amazonaws.com/Stage/"
