--- reInvent2017/tollcost/tollcost.yml
@@ -1,32 +1,0 @@
-AWSTemplateFormatVersion: '2010-09-09'
-Transform: 'AWS::Serverless-2016-10-31'
-
-Parameters:
-  HereAppId:
-    Type: String
-    Description: HERE Maps API App ID (https://developer.here.com/faqs#access-control)
-  HereAppCode:
-    Type: String
-    Description: HERE Maps API App Code (https://developer.here.com/faqs#access-control)
-
-Resources:
-  TollCostFunction:
-    Type: 'AWS::Serverless::Function'
-    Properties:
-      Runtime: nodejs6.10
-      Handler: tollcost.tollcostGET
-      Description: 'Calculate tolls for international driving routes.'
-      Environment:
-        Variables:
-          HERE_APP_ID: !Ref HereAppId
-          HERE_APP_CODE: !Ref HereAppCode
-      Events:
-        GET:
-          Type: Api
-          Properties:
-            Path: /tollcost
-            Method: get
-Outputs:
-  HereApiURL:
-    Description: "AWS API Gateway URL for Here.com Toll Cost API (CIT)."
-    Value: !Sub "https://${ServerlessRestApi}.execute-api.${AWS::Region}.amazonaws.com/Stage/"
