--- reInvent2017/traffic/traffic.yml
@@ -1,32 +1,0 @@
-AWSTemplateFormatVersion: '2010-09-09'
-Transform: 'AWS::Serverless-2016-10-31'
-
-Parameters:
-  HereAppId:
-    Type: String
-    Description: HERE Maps API App ID (https://developer.here.com/faqs#access-control)
-  HereAppCode:
-    Type: String
-    Description: HERE Maps API App Code (https://developer.here.com/faqs#access-control)
-
-Resources:
-  TrafficFunction:
-    Type: 'AWS::Serverless::Function'
-    Properties:
-      Runtime: nodejs6.10
-      Handler: traffic.trafficGET
-      Description: 'Display map tiles with real-time traffic flow information overlays reflected as colored lines drawn on affected streets and roads and get access to real-time traffic flow and incident data in XML and JSON.'
-      Environment:
-        Variables:
-          HERE_APP_ID: !Ref HereAppId
-          HERE_APP_CODE: !Ref HereAppCode
-      Events:
-        GET:
-          Type: Api
-          Properties:
-            Path: /traffic/{A}/{B}/{C}
-            Method: get
-Outputs:
-  HereApiURL:
-    Description: "AWS API Gateway URL for Here.com Traffic API (CIT)."
-    Value: !Sub "https://${ServerlessRestApi}.execute-api.${AWS::Region}.amazonaws.com/Stage/"
