--- reInvent2017/transit/transit.yml
@@ -1,32 +1,0 @@
-AWSTemplateFormatVersion: '2010-09-09'
-Transform: 'AWS::Serverless-2016-10-31'
-
-Parameters:
-  HereAppId:
-    Type: String
-    Description: HERE Maps API App ID (https://developer.here.com/faqs#access-control)
-  HereAppCode:
-    Type: String
-    Description: HERE Maps API App Code (https://developer.here.com/faqs#access-control)
-
-Resources:
-  TransitFunction:
-    Type: 'AWS::Serverless::Function'
-    Properties:
-      Runtime: nodejs6.10
-      Handler: transit.transitGET
-      Description: 'Get access to features that go beyond simple estimated transit routing like search, next departures, transit isoline and pedestrian access points.'
-      Environment:
-        Variables:
-          HERE_APP_ID: !Ref HereAppId
-          HERE_APP_CODE: !Ref HereAppCode
-      Events:
-        GET:
-          Type: Api
-          Properties:
-            Path: /transit
-            Method: get
-Outputs:
-  HereApiURL:
-    Description: "AWS API Gateway URL for Here.com Public Transit API (CIT)."
-    Value: !Sub "https://${ServerlessRestApi}.execute-api.${AWS::Region}.amazonaws.com/Stage/"
