+++ serverlessFunctions/mapimage/package.json
