+++ serverlessFunctions/maptile/maptile.yml
@@ -1,0 +1,77 @@
+AWSTemplateFormatVersion: '2010-09-09'
+Transform: AWS::Serverless-2016-10-31
+
+Parameters:
+  HereAppId:
+    Type: String
+    Description: HERE Location Services API App ID (https://saas.awsmarketplace.here.com/faq#api-access)
+  HereAppCode:
+    Type: String
+    Description: HERE Location Services API App Code (https://saas.awsmarketplace.here.com/faq#api-access)
+
+Resources:
+  HereApiMapTileGateway:
+    Type: AWS::Serverless::Api
+    Properties:
+      StageName: Prod
+      DefinitionBody:
+        swagger: "2.0"
+        info:
+          title:
+            Ref: AWS::StackName
+        paths:
+          "/maptile/{type}/{scheme}/{zoomlevel}/{col}/{row}/{rez}":
+            get:
+              produces:
+              - "application/json"
+              responses:
+                "200":
+                  description: "200 response"
+                  schema:
+                    $ref: "#/definitions/Empty"
+              x-amazon-apigateway-integration:
+                responses:
+                  default:
+                    statusCode: "200"
+                requestTemplates:
+                  application/json: "#set($allParams = $input.params())\n{\n  \"body-json\"\
+                        \ : $input.json('$'),\n  \"params\" : {\n    #foreach($type in $allParams.keySet())\n\
+                        \      #set($params = $allParams.get($type))\n      \"$type\" : {\n  \
+                        \      #foreach($paramName in $params.keySet())\n          \"$paramName\"\
+                        \ : \"$util.escapeJavaScript($params.get($paramName))\"\n          #if($foreach.hasNext),#end\n\
+                        \        #end\n      }\n    #if($foreach.hasNext),#end\n    #end\n  }\n\
+                        }"
+                uri: !Sub "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${MapTileFunction.Arn}/invocations"
+                passthroughBehavior: "never"
+                httpMethod: "POST"
+                contentHandling: "CONVERT_TO_TEXT"
+                type: "aws"
+        definitions:
+          Empty:
+            type: "object"
+            title: "Empty Schema"
+        x-amazon-apigateway-binary-media-types:
+        - "image/jpeg"
+  MapTileFunction:
+    Type: AWS::Serverless::Function
+    Properties:
+      Runtime: nodejs6.10
+      Handler: maptile.maptileGET
+      Description: 'A lambda function to show fresh day-time map tiles in multiple styles (e.g. base, aerial) including rendered live-traffic tiles for flow overlay.'
+      Environment:
+        Variables:
+          HERE_APP_ID: !Ref HereAppId
+          HERE_APP_CODE: !Ref HereAppCode
+      MemorySize: 512
+      Events:
+        GET:
+          Type: Api
+          Properties:
+            Path: /maptile/{type}/{scheme}/{zoomlevel}/{col}/{row}/{rez}
+            Method: GET
+            RestApiId:
+              Ref: HereApiMapTileGateway
+Outputs:
+  HereApiURL:
+    Description: "AWS API Gateway URL for HERE Map Tile API."
+    Value: !Sub "https://${HereApiMapTileGateway}.execute-api.${AWS::Region}.amazonaws.com/Prod/"
