+++ serverlessFunctions/weather/weather.yml
@@ -1,0 +1,32 @@
+AWSTemplateFormatVersion: '2010-09-09'
+Transform: 'AWS::Serverless-2016-10-31'
+
+Parameters:
+  HereAppId:
+    Type: String
+    Description: HERE Location Services API App ID (https://saas.awsmarketplace.here.com/faq#api-access)
+  HereAppCode:
+    Type: String
+    Description: HERE Location Services API App Code (https://saas.awsmarketplace.here.com/faq#api-access)
+
+Resources:
+  WeatherFunction:
+    Type: 'AWS::Serverless::Function'
+    Properties:
+      Runtime: nodejs6.10
+      Handler: weather.weatherGET
+      Description: 'A lambda function to give insights into real-time weather forecasts, alerts, and astronomical info for any location.'
+      Environment:
+        Variables:
+          HERE_APP_ID: !Ref HereAppId
+          HERE_APP_CODE: !Ref HereAppCode
+      Events:
+        GET:
+          Type: Api
+          Properties:
+            Path: /weather
+            Method: get
+Outputs:
+  HereApiURL:
+    Description: "AWS API Gateway URL for HERE Weather API."
+    Value: !Sub "https://${ServerlessRestApi}.execute-api.${AWS::Region}.amazonaws.com/Prod/"
